requirements.txt
toml~=0.10.0
pip~=19.2.3
py~=1.8.1
lxml~=4.5.0
setuptools~=41.2.0
aiogram~=2.9.2
python-dotenv~=0.15.0
requests~=2.25.0
bs4~=0.0.1
beautifulsoup4~=4.9.1